test: refactor - move asyncio_run helper method to utils/__init__.py
